+++ src/snake/snakeview.cpp
@@ -117,7 +117,7 @@ void SnakeView::gameOver()
     update();
     refresh();
 
-    QTimer::singleShot(200, this, [this, newBest] {
+    announceLater(200, [this, newBest] {
         QMessageBox box(this);
         box.setWindowTitle(QStringLiteral("Game over"));
         box.setText(QStringLiteral("The snake stopped."));